Use our own customed DefaultHttpClient for CF
With android version older than 2.3.1 the default keystore doesn't have
information about the domain SSL setup on CloudFiles UK.

We use our own customed DefaultHttpClient which use the android 2.3.1
bks instead of the global keystore available on the system.

Patch is bigger than it should be since we need to pass the Activity
context around classes.